make pointers in solution variables

+++ include/constants.h
@@ -1,7 +1,7 @@
 #ifndef constants_h__
 #define constants_h__
 // Earth constants
-const double R{6378245.0};
-const double U{0.0000726}; // ((Pi * Wspd_earth)) / 180 degree
-const double G{9.81};
+const float R{6378245.0};
+const float U{0.0000726}; // ((Pi * Wspd_earth)) / 180 degree
+const float G{9.81};
 #endif
